pyproject.toml
[project]
name = "blurifier"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "better-profanity>=0.7.0",
    "celery[redis]>=5.5.3",
    "django-debug-toolbar>=6.0.0",
    "django>=5.2",
    "django-celery-beat>=2.8.1",
    "django-ninja>=1.4.3",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.0",
    "flower>=2.0.1",
    "gunicorn>=23.0.0",
    "jinja2>=3.1.6",
    "psycopg[binary,pool]>=3.2.9",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "redis>=5.2.1",
    "uvicorn>=0.35.0",
    "elasticsearch[async]>=9.1.0",
]
[project.optional-dependencies]
dev = [
    "fakeredis>=2.30.1",
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-django>=4.11.1",
    "ruff>=0.12.5",
]
[tool.ruff.format]
quote-style = "single"
